src/requirements.txt
paho_mqtt==2.1.0
SQLAlchemy==2.0.44
mysql-connector-python==9.5.0